Move generator classes into new files
