+++ lib/ffi_gen/generator/define.rb
@@ -1,0 +1,14 @@
+module FFIGen
+  class Generator
+    class Define
+
+      def initialize(generator, name, parameters, value)
+        @generator = generator
+        @name = name
+        @parameters = parameters
+        @value = value
+      end
+
+    end
+  end
+end
