+++ lib/ffi_gen/generator/string_type.rb
@@ -1,0 +1,11 @@
+module FFIGen
+  class Generator
+    class StringType < Type
+
+      def name
+        Name.new ["string"]
+      end
+
+    end
+  end
+end
