+++ lib/ffi_gen/generator/type.rb
@@ -1,0 +1,7 @@
+module FFIGen
+  class Generator
+    class Type
+
+    end
+  end
+end
