+++ lib/ffi_gen/generator/unknown_type.rb
@@ -1,0 +1,11 @@
+module FFIGen
+  class Generator
+    class UnknownType < Type
+
+      def name
+        Name.new ["unknown"]
+      end
+
+    end
+  end
+end
